// src/app/models/maff.model.ts
export class Maff {
    idMaff: number;
    numeScoala: string;
    numeU: string;
    judetS: string;
    localitateS: string;
    localitateU: string;
    statutU: string;
    scoalaId: number;
    existaClasa15: boolean;
    romiScoala: number;
    parintiScoala: number;
    dizabilitatiScoala: number;
    raspuns38: string;
    raspuns39: string;
    raspuns40: string;
    raspuns41: string;
    raspuns46: string;
    raspuns47: string;
    romiP: number;
    romiLocalitate: number;
    badClasaNume: string[];
    badClasaId: number[];
    total: number;
    romi: number;
    dizabilitati: number;
    parinti: number;
    burse: number;
    repetenti: number;
    online: number;
    remediala: number;
    totalC: number;
    romiC: number;
    dizabilitatiC: number;
    parintiC: number;
    burseC: number;
    repetentiC: number;
    onlineC: number;
    remedialaC: number;
    alerta1: string;
    alerta2: string;
    alerta3: string;
    alerta4: string;
    alerta5: string;
    alerta6: string;
    alerta7: string;
    alerta8: string;
    alerta9: string;
    alerta10: string;
    alerta11: string;
    nrDep: string;
  
    constructor(input?: any) {
      Object.assign(this, input);
    }
  }
  